Extract logic for better readability
